Make git-cookie-authdaemon more robust.
Once the daemon has backgrounded itself, retry fetching the cookie if
there's an error, rather than just allowing the daemon to crash. Also,
throttle the minimum interval between new attempts to fetch the cookie,
as otherwise if the server returns a cookie that expires in less than
REFRESH seconds, the script will keep trying to refresh the cookie as
fast as it possibly can until it gets a new one, which triggers DoS
prevention limits.

Failures on the initial cookie fetch are not retried, as the daemon is
still in the foreground at that point and hopefully either a human or a
logging process is watching it.